feat: Add hook to override default options (FOR TESTS ONLY)
In order to re-use existing test code for different FileCopyMethod
implementations, we need some way to change which one the tests
implicitly use (AKA: the default).

Rather than wrapping the `Copy` function to inject the option or
giving users the capability to change the default, we give the test
code a hook for doing what it needs to do.